--- a/docs/source/paper/paper.rst
+++ b/docs/source/paper/paper.rst
@@ -1,7 +1,0 @@
-.. _paperrst:
-
-==================
-RST-Rendered Paper
-==================
-
-.. mdinclude:: ../../../joss/paper.md